Code cleanup

+++ mud.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <unistd.h>
+#include <stddef.h>
 
 struct mud;
 
@@ -13,5 +13,5 @@ int mud_peer (struct mud *, const char *, const char *);
 int mud_pull (struct mud *);
 int mud_push (struct mud *);
 
-ssize_t mud_recv (struct mud *, void *, size_t);
-ssize_t mud_send (struct mud *, const void *, size_t);
+int mud_recv (struct mud *, void *, size_t);
+int mud_send (struct mud *, const void *, size_t);
